Bug fixes
Need to generate a brain mask in ASL-gridded T1 space for passing to
oxford_asl later in the pipeline.

Apply motion and distortion correction to entire ASL series, not just
the first volume.

Also need to apply the same transformations to the scaling factors.

Use the corrected (bias and MT) calibration image as a target for
registration and apply the motion/distortion correction to this, not the
original calibration image, for use in oxford_asl later in the
pipeline.