New fix for conversion from .dat to .mat
bbregister's conversion to fslmat occasionally gives a final row != [0 0 0 1] due to what looks like a numerical precision issue.
lta_convert had seemed to be robust to this on a few test cases but doesn't always work.
Here, we instead round entries in the matrix to 5 d.p. if the final row != [0 0 0 1].