Move create_dirs to utils.py
create_dirs already existed in utils.py but I was still using the still-existing version in initial_bookkeeping. I've now deleted this old version and switched to using the version in utils.py.

(Both versions are the same but it makes more sense for it to be in utils.py.)