Tidy asl_correction
A mix of fsl.data.image.Image and nibabel.nifti1.Nifti1Image was being used. I've changed so that only the latter is being used to be consistent.

A number of files were being located via the json. This has been updated so that they're found relative to parent directories (e.g. the ASL sequence) or provided as an argument (e.g. the bias field and the corrected calibration image).